Action_Get to Vuzix
