+++ Phone/app/src/main/java/com/facilitation/phone/MainActivity.kt
@@ -1,6 +1,5 @@
 package com.facilitation.phone
 
-import android.content.Intent
 import android.os.Bundle
 import com.google.android.material.bottomnavigation.BottomNavigationView
 import androidx.appcompat.app.AppCompatActivity
@@ -9,12 +8,10 @@ import androidx.navigation.ui.AppBarConfiguration
 import androidx.navigation.ui.setupActionBarWithNavController
 import androidx.navigation.ui.setupWithNavController
 import com.facilitation.phone.databinding.ActivityMainBinding
-import com.vuzix.connectivity.sdk.Connectivity
 
 class MainActivity : AppCompatActivity() {
 
     private lateinit var binding: ActivityMainBinding
-    var ACTION_GET = "com.facilitation.vuzix.GET"
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -34,19 +31,5 @@ class MainActivity : AppCompatActivity() {
         )
         setupActionBarWithNavController(navController, appBarConfiguration)
         navView.setupWithNavController(navController)
-    }
-
-    private fun getRemoteDeviceModelClicked() {
-        try {
-            val connectivity = Connectivity.get(this)
-            val device = connectivity.device
-            if (device != null) {
-                val getIntent = Intent(this.ACTION_GET)
-                connectivity.sendBroadcast(getIntent)
-            }
-        }
-        catch (e:Exception) {
-            print(e.stackTrace)
-        }
     }
 }
